requirements.txt
pytest==8.3.3
fastapi==0.115.0
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
email-validator==2.2.0
dnspython==2.7.0
sqlalchemy==2.0.36
redis==5.2.1
rq==1.16.2
prometheus-client==0.21.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
hypothesis==6.112.0
